Module re_types_builder::codegen::cpp

Modules

array_builder ğŸ”’

forward_decl ğŸ”’

includes ğŸ”’

method ğŸ”’

Structs

CppCodeGenerator

QuotedObject ğŸ”’

Constants

ANGLE_BRACKET_LEFT_TOKEN ğŸ”’

ANGLE_BRACKET_RIGHT_TOKEN ğŸ”’

DOC_COMMENT_PREFIX_TOKEN ğŸ”’

DOC_COMMENT_SUFFIX_TOKEN ğŸ”’

HEADER_EXTENSION_TOKEN ğŸ”’

NEWLINE_TOKEN ğŸ”’

NORMAL_COMMENT_PREFIX_TOKEN ğŸ”’

NORMAL_COMMENT_SUFFIX_TOKEN ğŸ”’

Functions

archetype_as_component_batches ğŸ”’

archetype_indicator ğŸ”’

are_types_disjoint ğŸ”’

arrow_data_type_method ğŸ”’

component_to_data_cell_method ğŸ”’

fill_arrow_array_builder_method ğŸ”’

format_code ğŸ”’

generate_hpp_cpp ğŸ”’

hpp_type_extensions ğŸ”’

Retrieves code from an extension cpp file that should go to the generated header.

lines_from_docs ğŸ”’

new_arrow_array_builder_method ğŸ”’

pragma_once ğŸ”’

quote_append_field_to_builder ğŸ”’

quote_append_single_field_to_builder ğŸ”’

quote_append_single_value_to_builder ğŸ”’

Appends a single value to an arrow array builder.

quote_arrow_data_type ğŸ”’

quote_arrow_elem_type ğŸ”’

quote_arrow_field_type ğŸ”’

quote_comment ğŸ”’

quote_constants_header_and_cpp ğŸ”’

quote_doc_comment ğŸ”’

quote_doc_lines ğŸ”’

quote_element_type ğŸ”’

quote_field_docs ğŸ”’

quote_field_ptr_access ğŸ”’

quote_fill_arrow_array_builder ğŸ”’

quote_fqname_as_type_path ğŸ”’

quote_integer ğŸ”’

quote_num_items_per_value ğŸ”’

quote_obj_docs ğŸ”’

quote_variable ğŸ”’

quote_variable_with_docstring ğŸ”’

single_field_constructor_methods ğŸ”’

static_constructor_for_enum_type ğŸ”’

e.g. static Angle radians(float radians); -> auto angle = Angle::radians(radians);

string_from_token_stream ğŸ”’