Module re_types_builder::codegen::docs::snippets_ref

Generate the snippets reference.

Structs

Config 🔒

See docs/snippets/snippets.toml for more info

KnownObjects 🔒

Neatly organized Objects (archetypes, components, etc).

OptOut 🔒

Snippet 🔒

Everything we know about a snippet, including which objects (archetypes, components, etc) it references.

Snippets 🔒

Maps objects (archetypes, components, etc) back to snippets.

SnippetsRef 🔒

SnippetsRefCodeGenerator

Functions

collect_snippets_recursively 🔒

is_speculative 🔒

Returns true if the given name has not been released yet.

make_speculative_if_needed 🔒

Appends ?speculative-link to the given link if the associated object has not been released yet.