Module re_types_builder::codegen::python

Implements the Python codegen pass.

Structs

PythonCodeGenerator

QuotedObject ğŸ”’

Statics

ONCE ğŸ”’

__SEEN_MESSAGES ğŸ”’

Traits

PythonObjectExt ğŸ”’

Python-specific helpers for Object.

Functions

f ğŸ”’

f ğŸ”’

f ğŸ”’

f ğŸ”’

f ğŸ”’

format_python ğŸ”’

lib_source_code ğŸ”’

load_overrides ğŸ”’

Inspect _overrides sub-packages for manual override of the generated code.

python_project_path ğŸ”’

quote_aliases_from_object ğŸ”’

Only applies to datatypes and components.

quote_array_method_from_obj ğŸ”’

Automatically implement __array__ if the object is a single npt.ArrayLike/integer/floating-point field.

quote_arrow_datatype ğŸ”’

quote_arrow_field ğŸ”’

quote_arrow_support_from_delegating_component ğŸ”’

quote_arrow_support_from_obj ğŸ”’

Arrow support objects

quote_doc_from_docs ğŸ”’

quote_doc_from_fields ğŸ”’

quote_field_converter_from_field ğŸ”’

Returns a default converter function for the given field.

quote_field_type_from_field ğŸ”’

Returns type name as string and whether it was force unwrapped.

quote_import_clauses_from_field ğŸ”’

quote_manifest ğŸ”’

quote_metadata_map ğŸ”’

quote_native_types_method_from_obj ğŸ”’

Automatically implement __str__, __int__, or __float__ method if the object has a single field of the corresponding type that is not optional.

quote_objects ğŸ”’

Returns all filepaths + all object names.

quote_type_from_element_type ğŸ”’

quote_union_aliases_from_object ğŸ”’

Quote typing aliases for union datatypes. The types for the union arms are automatically included.

run_black ğŸ”’

run_ruff ğŸ”’

write_file ğŸ”’

write_files ğŸ”’