Function re_types_builder::codegen::python::quote_field_type_from_field

fn quote_field_type_from_field(
    _objects: &Objects,
    field: &ObjectField,
    unwrap: bool,
) -> (String, bool)

Returns type name as string and whether it was force unwrapped.

Specifying unwrap = true will unwrap the final type before returning it, e.g. Vec<String> becomes just String. The returned boolean indicates whether there was anything to unwrap at all.