re_types_builder::codegen::python
Trait PythonObjectExt

trait PythonObjectExt {
    // Required methods
    fn is_delegating_component(&self) -> bool;
    fn is_non_delegating_component(&self) -> bool;
    fn delegate_datatype<'a>(&self, objects: &'a Objects) -> Option<&'a Object>;
}

Python-specific helpers for Object.

Required Methods

fn is_delegating_component(&self) -> bool

Returns true if the object is a delegating component.

Components can either use a native type, or a custom datatype. In the latter case, the component delegates its implementation to the datatype.

fn is_non_delegating_component(&self) -> bool

Returns true if the object is a non-delegating component.

fn delegate_datatype<'a>(&self, objects: &'a Objects) -> Option<&'a Object>

If the object is a delegating component, returns the datatype it delegates to.