re_types_builder::codegen::rust::deserializer
Function quote_arrow_deserializer

pub fn quote_arrow_deserializer(
    type_registry: &TypeRegistry,
    objects: &Objects,
    obj: &Object,
) -> TokenStream

This generates code that deserializes a runtime Arrow payload into the specified obj, taking Arrow-transparency into account.

This short-circuits on error using the try (?) operator: the outer scope must be one that returns a Result<_, DeserializationError>!

There is a 1:1 relationship between quote_arrow_deserializer and Loggable::from_arrow_opt:

fn from_arrow_opt(data: &dyn ::arrow::array::Array) -> DeserializationResult<Vec<Option<Self>>> {
    Ok(#quoted_deserializer)
}

This tells you two things:

The runtime Arrow payload is always held in a variable data, identified as data_src below.

The returned TokenStream must always instantiates a Vec<Option<Self>>.

Performance vs validation

The deserializers are designed for maximum performance, assuming the incoming data is correct. If the data is not correct, the deserializers will return an error, but never panic or crash.

TODO(#5305): Currently we’re doing a lot of checking for exact matches. We should instead assume data is correct and handle errors gracefully.

Understanding datatypes

There are three (!) datatypes involved in the deserialization process:

The object’s native Rust type, which was derived from its IDL definition by the codegen framework.

The object’s Arrow datatype, which was also derived from its IDL definition.

The runtime payload’s advertised Arrow datatype.

The deserialization process is entirely driven by our own compile-time IDL-derived definitions: the runtime payload’s advertised Arrow datatype is only ever used as a mean of checking whether the data we receive can be coerced one way or another into something that fit our schema.

In some places that coercion can be very strict (if the data doesn’t match exactly, we abort with a runtime error) while in other it might be more relaxed for performance reasons (e.g. ignore the fact that the data has a bitmap altogether).