re_types_builder::codegen::rust::deserializer
Function quote_iterator_transparency

fn quote_iterator_transparency(
    objects: &Objects,
    datatype: &DataType,
    iter_kind: IteratorKind,
    extra_wrapper: Option<TokenStream>,
) -> TokenStream

This generates code that maps the data in an iterator in order to apply the Arrow transparency rules to it, if necessary.

This can often become a very difficult job due to all the affixes that might be involved: fallibility, nullability, transparency, tuple structs… This function will just do the right thing.

If extra_wrapper is specified, this will also wrap the resulting data in $extra_wrapper(data).

Have a look around in this file for examples of use.