Function re_types_builder::codegen::rust::serializer::quote_arrow_field_serializer

fn quote_arrow_field_serializer(
    objects: &Objects,
    datatype: &DataType,
    quoted_datatype: &dyn ToTokens,
    obj_field: Option<&ObjectField>,
    bitmap_src: &Ident,
    elements_are_nullable: bool,
    data_src: &Ident,
    inner_repr: InnerRepr
) -> TokenStream

Writes out code to serialize a single field.

If elements_are_nullable is false, then we ignore null elements in the input data. This is useful for:

unions: nullability is encoded as a separate variant

lists inside of fields that are lists: we don’t support intra-list nullability TODO(#2993): However, we still emit a validity bitmap for lists inside lists to validity_bitmap_identsince Python and Rust do so.