Module re_types_builder::codegen::rust::to_archetype

Generates code in re_query_cache2 so that cached results can easily be converted to ready-to-use archetypes.

That code needs to be generated directly in the caching crates as it needs access to the cached queries and results as well as the promise resolving machinery. Generating such code in the usual places would result in one giant cycle dependency chain.

Constants

COMMENT_RECOMMENDED_OPTIONAL_TOKEN 🔒

COMMENT_REQUIRED_TOKEN 🔒

COMMENT_SEPARATOR_TOKEN 🔒

NEWLINE_TOKEN 🔒

Functions

generate_impls 🔒

generate_mod 🔒

generate_to_archetype_impls

quote_fqname_as_type_path 🔒

quote_to_archetype_impl 🔒