Trait re_types_builder::codegen::CodeGenerator

pub trait CodeGenerator {
    // Required method
    fn generate(
        &mut self,
        reporter: &Reporter,
        objects: &Objects,
        arrow_registry: &ArrowRegistry
    ) -> GeneratedFiles;
}

Implements the codegen pass.

Required Methods

fn generate( &mut self, reporter: &Reporter, objects: &Objects, arrow_registry: &ArrowRegistry ) -> GeneratedFiles

Generates user-facing code from crate::Objects.

Returns the expected paths and contents of all generated files. It is the responsibility of the caller to actually do something with that data (e.g. write them to disk).