re_types_builder::data_type
Type Alias LazyField

pub type LazyField = GenericField<LazyDatatype>;

A yet-to-be-resolved Field.

Type resolution is a two-pass process as we first need to register all existing types before we can denormalize their definitions into their parents.

Aliased Type

struct LazyField {
    pub name: String,
    pub data_type: LazyDatatype,
    pub is_nullable: bool,
    pub metadata: BTreeMap<String, String>,
}

Fields
name: String

Its name

data_type: LazyDatatype

Its logical DataType

is_nullable: bool

Its nullability

metadata: BTreeMap<String, String>

Additional custom (opaque) metadata.

Implementations

impl LazyField

pub fn resolve(&self, registry: &TypeRegistry) -> Field