Module re_types_builder::docs::doclink_translation

We support doclinks in our docstrings.

They need to follow this format:

/// See also [archetype.Image].
table Tensor { … }

This module is all about translating these doclinks to the different Targets.

The code is not very efficient, but it is simple and works.

Functions

tokenize 🔒

tokenize_with 🔒

translate_doc_line

Convert Rerun-style doclinks to the target language.

translate_doclink 🔒

try_translate_doclink 🔒