re_types_builder
Function compile_binary_schemas

pub fn compile_binary_schemas(
    include_dir_path: impl AsRef<Utf8Path>,
    output_dir_path: impl AsRef<Utf8Path>,
    entrypoint_path: impl AsRef<Utf8Path>,
)

Compiles binary reflection dumps from flatbuffers definitions.

Requires flatc available in $PATH.

Panics on error.

include_dir_path: path to the root directory of the fbs definition tree.

output_dir_path: output directory, where the binary schemas will be stored.

entrypoint_path: path to the root file of the fbs definition tree.