Function re_types_builder::generate_lang_agnostic

pub fn generate_lang_agnostic(
    reporter: &Reporter,
    include_dir_path: impl AsRef<Utf8Path>,
    entrypoint_path: impl AsRef<Utf8Path>
) -> (Objects, ArrowRegistry)

Handles the first 3 language-agnostic passes of the codegen pipeline:

Generate binary reflection dumps for our definitions.

Run the semantic pass

Compute the Arrow registry

Panics on error.

include_dir_path: path to the root directory of the fbs definition tree.

entrypoint_path: path to the root file of the fbs definition tree.