Struct re_types_builder::objects::Object

pub struct Object {
    pub virtpath: String,
    pub filepath: Utf8PathBuf,
    pub fqname: String,
    pub pkg_name: String,
    pub name: String,
    pub docs: Docs,
    pub kind: ObjectKind,
    pub attrs: Attributes,
    pub fields: Vec<ObjectField>,
    pub specifics: ObjectSpecifics,
    pub datatype: Option<LazyDatatype>,
}

A high-level representation of a flatbuffers object, which can be either a struct, a union or an enum.

Fields
virtpath: String

Utf8Path of the associated fbs definition in the Flatbuffers hierarchy, e.g. //rerun/components/point2d.fbs.

filepath: Utf8PathBuf

Absolute filepath of the associated fbs definition.

fqname: String

Fully-qualified name of the object, e.g. rerun.components.Position2D.

pkg_name: String

Fully-qualified package name of the object, e.g. rerun.components.

name: String

PascalCase name of the object type, e.g. Position2D.

docs: Docs

The object’s multiple layers of documentation.

kind: ObjectKind

The object’s kind: datatype, component or archetype.

attrs: Attributes

The object’s attributes.

fields: Vec<ObjectField>

The object’s inner fields, which can be either struct members or union values.

These are pre-sorted, in ascending order, using their order attribute.

specifics: ObjectSpecifics

Properties that only apply to either structs or unions.

datatype: Option<LazyDatatype>

The Arrow datatype of this Object, or None if the object is Arrow-transparent.

This is lazily computed when the parent object gets registered into the Arrow registry and will be None until then.

Implementations

impl Object

pub fn from_raw_object( include_dir_path: impl AsRef<Utf8Path>, enums: &[FbsEnum<'_>], objs: &[FbsObject<'_>], obj: &FbsObject<'_> ) -> Self

Resolves a raw crate::Object into a higher-level representation that can be easily interpreted and manipulated.

pub fn from_raw_enum( include_dir_path: impl AsRef<Utf8Path>, enums: &[FbsEnum<'_>], objs: &[FbsObject<'_>], enm: &FbsEnum<'_> ) -> Self

Resolves a raw FbsEnum into a higher-level representation that can be easily interpreted and manipulated.

pub fn get_attr<T>(&self, name: impl AsRef<str>) -> Twhere T: FromStr, T::Err: Error + Send + Sync + 'static,

pub fn try_get_attr<T>(&self, name: impl AsRef<str>) -> Option<T>where T: FromStr, T::Err: Error + Send + Sync + 'static,

pub fn is_attr_set(&self, name: impl AsRef<str>) -> bool

pub fn is_struct(&self) -> bool

pub fn is_enum(&self) -> bool

pub fn is_union(&self) -> bool

pub fn is_arrow_transparent(&self) -> bool

fn is_transparent(&self) -> bool

pub fn has_default_destructor(&self, objects: &Objects) -> bool

Is the destructor trivial/default (i.e. is this simple data with no allocations)?

pub fn relative_filepath(&self) -> Option<&Utf8Path>

Try to find the relative file path of the .fbs source file.

pub fn snake_case_name(&self) -> String

The snake_case name of the object, e.g. translation_and_mat3x3.

pub fn is_testing(&self) -> bool

Returns true if this object is part of testing and not to be used in the production SDK.