Struct re_types_builder::objects::Objects

pub struct Objects {
    pub objects: BTreeMap<String, Object>,
}

The result of the semantic pass: an intermediate representation of all available object types; including structs, enums and unions.

Fields
objects: BTreeMap<String, Object>

Maps fully-qualified type names to their resolved object definitions.

Implementations

impl Objects

pub fn from_buf(include_dir_path: impl AsRef<Utf8Path>, buf: &[u8]) -> Self

Runs the semantic pass on a serialized flatbuffers schema.

The buffer must be a serialized FbsSchema (i.e. .bfbs data).

pub fn from_raw_schema( include_dir_path: impl AsRef<Utf8Path>, schema: &FbsSchema<'_> ) -> Self

Runs the semantic pass on a deserialized flatbuffers FbsSchema.

impl Objects

pub fn ordered_objects_mut( &mut self, kind: Option<ObjectKind> ) -> Vec<&mut Object>

Returns all available objects, pre-sorted in ascending order based on their order attribute.

pub fn ordered_objects(&self, kind: Option<ObjectKind>) -> Vec<&Object>

Returns all available objects, pre-sorted in ascending order based on their order attribute.