Struct re_types_builder::ObjectField

pub struct ObjectField {
    pub virtpath: String,
    pub filepath: Utf8PathBuf,
    pub fqname: String,
    pub pkg_name: String,
    pub name: String,
    pub docs: Docs,
    pub typ: Type,
    pub attrs: Attributes,
    pub order: u32,
    pub is_nullable: bool,
    pub datatype: Option<LazyDatatype>,
}

A high-level representation of a flatbuffers field, which can be either a struct member or a union value.

Fields
virtpath: String

Utf8Path of the associated fbs definition in the Flatbuffers hierarchy, e.g. //rerun/components/point2d.fbs.

filepath: Utf8PathBuf

Absolute filepath of the associated fbs definition.

fqname: String

Fully-qualified name of the field, e.g. rerun.components.Position2D#position.

pkg_name: String

Fully-qualified package name of the field, e.g. rerun.components.

name: String

Name of the field, e.g. x.

For struct fields this is usually snake_case, but for enums it is usually PascalCase.

docs: Docs

The field’s multiple layers of documentation.

typ: Type

The field’s type.

attrs: Attributes

The field’s attributes.

order: u32

The struct field’s order attribute’s value, which is mandatory for struct fields (otherwise their order is undefined).

is_nullable: bool

Whether the field is nullable.

datatype: Option<LazyDatatype>

The Arrow datatype of this ObjectField.

This is lazily computed when the parent object gets registered into the Arrow registry and will be None until then.

Implementations

impl ObjectField

pub fn from_raw_object_field( reporter: &Reporter, include_dir_path: impl AsRef<Utf8Path>, enums: &[FbsEnum<'_>], objs: &[FbsObject<'_>], obj: &FbsObject<'_>, field: &FbsField<'_> ) -> Self

pub fn from_raw_enum_value( reporter: &Reporter, include_dir_path: impl AsRef<Utf8Path>, enums: &[FbsEnum<'_>], objs: &[FbsObject<'_>], enm: &FbsEnum<'_>, val: &FbsEnumVal<'_> ) -> Self

fn is_transparent(&self) -> bool

pub fn get_attr<T>(&self, name: impl AsRef<str>) -> T
where T: FromStr, T::Err: Error + Send + Sync + 'static,

pub fn try_get_attr<T>(&self, name: impl AsRef<str>) -> Option<T>
where T: FromStr, T::Err: Error + Send + Sync + 'static,

pub fn has_attr(&self, name: impl AsRef<str>) -> bool

pub fn snake_case_name(&self) -> String

The snake_case name of the field, e.g. translation_and_mat3x3.

pub fn pascal_case_name(&self) -> String

The PascalCase name of the field, e.g. TranslationAndMat3x3.

pub fn is_testing(&self) -> bool

Returns true if this object is part of testing and not to be used in the production SDK.

pub fn kind(&self) -> Option<FieldKind>