re_types_builder
Struct Objects

pub struct Objects {
    pub objects: BTreeMap<String, Object>,
}

The result of the semantic pass: an intermediate representation of all available object types; including structs, enums and unions.

Fields
objects: BTreeMap<String, Object>

Maps fully-qualified type names to their resolved object definitions.

Implementations

impl Objects

pub fn from_buf( reporter: &Reporter, include_dir_path: impl AsRef<Utf8Path>, buf: &[u8], ) -> Self

Runs the semantic pass on a serialized flatbuffers schema.

The buffer must be a serialized FbsSchema (i.e. .bfbs data).

pub fn from_raw_schema( reporter: &Reporter, include_dir_path: impl AsRef<Utf8Path>, schema: &FbsSchema<'_>, ) -> Self

Runs the semantic pass on a deserialized flatbuffers FbsSchema.

impl Objects

pub fn get(&self, fqname: &str) -> Option<&Object>

pub fn values(&self) -> impl Iterator<Item = &Object>

pub fn objects_of_kind(&self, kind: ObjectKind) -> impl Iterator<Item = &Object>

Returns all available objects of the given kind.