re_types_builder::codegen
Module common

Helpers common to all codegen passes.

Structs

Example

ExampleInfo

ImageStack

RerunImageUrl

SnippetId

Enums

ImageUrl

Traits

StringExt

Functions

collect_snippets_for_api_docs

is_blank ๐

remove_orphaned_files

Remove orphaned files in all directories present in files.

write_file

Write file if any changes were made and ensure folder hierarchy exists.