Module re_types_builder::codegen::cpp

Modules

array_builder ğŸ”’

forward_decl ğŸ”’

includes ğŸ”’

method ğŸ”’

Structs

CppCodeGenerator

QuotedObject ğŸ”’

Constants

ANGLE_BRACKET_LEFT_TOKEN ğŸ”’

ANGLE_BRACKET_RIGHT_TOKEN ğŸ”’

DOC_COMMENT_PREFIX_TOKEN ğŸ”’

DOC_COMMENT_SUFFIX_TOKEN ğŸ”’

HEADER_EXTENSION_TOKEN ğŸ”’

NEWLINE_TOKEN ğŸ”’

NORMAL_COMMENT_PREFIX_TOKEN ğŸ”’

NORMAL_COMMENT_SUFFIX_TOKEN ğŸ”’

Traits

CppObjectExtensions ğŸ”’

Functions

add_copy_assignment_and_constructor ğŸ”’

archetype_serialize ğŸ”’

are_types_disjoint ğŸ”’

arrow_data_type_method ğŸ”’

fill_arrow_array_builder_method ğŸ”’

generate_hpp_cpp ğŸ”’

generate_object_files ğŸ”’

hpp_type_extensions ğŸ”’

Retrieves code from an extension cpp file that should go to the generated header.

lines_from_docs ğŸ”’

pragma_once ğŸ”’

quote_append_field_to_builder ğŸ”’

quote_append_single_field_to_builder ğŸ”’

quote_append_single_value_to_builder ğŸ”’

Appends a single value to an arrow array builder.

quote_archetype_field_type ğŸ”’

quote_arrow_data_type ğŸ”’

quote_arrow_elem_type ğŸ”’

quote_arrow_field_type ğŸ”’

quote_comment ğŸ”’

quote_deprecation_notice ğŸ”’

quote_doc_comment ğŸ”’

quote_doc_lines ğŸ”’

quote_element_type ğŸ”’

quote_field_docs ğŸ”’

quote_field_ptr_access ğŸ”’

quote_field_type ğŸ”’

quote_fill_arrow_array_builder ğŸ”’

quote_fqname_as_type_path ğŸ”’

quote_hide_from_docs ğŸ”’

quote_integer ğŸ”’

quote_loggable_hpp_and_cpp ğŸ”’

quote_num_items_per_value ğŸ”’

quote_obj_docs ğŸ”’

quote_variable ğŸ”’

quote_variable_with_docstring ğŸ”’

single_field_constructor_methods ğŸ”’

static_constructor_for_enum_type ğŸ”’

e.g. static Angle radians(float radians); -> auto angle = Angle::radians(radians);

string_from_token_stream ğŸ”’

to_arrow_method ğŸ”’

transparent_forwarded_fqname ğŸ”’

If the type forwards to another rerun defined type, returns the fully qualified name of that type.

Type Aliases

Result ğŸ”’