Module re_types_builder::codegen::python

Implements the Python codegen pass.

Modules

views 🔒

Structs

ExtensionClass 🔒

ExtensionClass represents an optional means of extending the generated python code

PythonCodeGenerator

Constants

ARRAY_METHOD 🔒

The standard numpy interface for converting to an array type

DEFERRED_PATCH_CLASS_METHOD 🔒

The method used for deferred patch class init. Use this for initialization constants that need to know the child (non-extension) class.

FIELD_CONVERTER_SUFFIX 🔒

The common suffix for method used to convert fields to their canonical representation.

INIT_METHOD 🔒

The standard python init method.

NATIVE_TO_PA_ARRAY_METHOD 🔒

The method used to convert a native type into a pyarrow array

Traits

PythonObjectExt 🔒

Python-specific helpers for Object.

Functions

code_for_enum 🔒

code_for_struct 🔒

code_for_union 🔒

fqname_to_type 🔒

lib_source_code 🔒

lines_from_docs 🔒

np_dtype_from_type 🔒

quote_aliases_from_object 🔒

Only applies to datatypes and components.

quote_array_method_from_obj 🔒

Automatically implement __array__ if the object is a single npt.ArrayLike/integer/floating-point field.

quote_arrow_datatype 🔒

quote_arrow_field 🔒

quote_arrow_serialization 🔒

Only implemented for some cases.

quote_arrow_support_from_obj 🔒

Arrow support objects

quote_clear_methods 🔒

quote_doc_from_fields 🔒

quote_doc_lines 🔒

Ends with double newlines, unless empty.

quote_examples 🔒

quote_field_converter_from_field 🔒

Returns a default converter function for the given field.

quote_field_type_from_field 🔒

Returns type name as string and whether it was force unwrapped.

quote_import_clauses_from_field 🔒

quote_import_clauses_from_fqname 🔒

quote_init_method 🔒

quote_init_parameter_from_field 🔒

quote_local_batch_type_imports 🔒

quote_manifest 🔒

quote_metadata_map 🔒

quote_native_types_method_from_obj 🔒

Automatically implement __str__, __int__, or __float__ as well as __hash__ methods if the object has a single field of the corresponding type that is not optional.

quote_obj_docs 🔒

Ends with double newlines, unless empty.

quote_parameter_type_alias 🔒

quote_type_from_element_type 🔒

quote_type_from_type 🔒

quote_union_aliases_from_object 🔒

Quote typing aliases for union datatypes. The types for the union arms are automatically included.

quote_union_kind_from_fields 🔒

write_init_file 🔒