re_types_builder::codegen::rust
Module deserializer

Enums

InnerRepr 🔒

IteratorKind 🔒

Functions

quote_array_downcast 🔒

Generates tokens that downcast the runtime Arrow array identifier by arr as cast_as, making sure to inject proper error handling.

quote_arrow_deserializer

This generates code that deserializes a runtime Arrow payload into the specified obj, taking Arrow-transparency into account.

quote_arrow_deserializer_buffer_slice

This generates code that deserializes a runtime Arrow payload into the specified obj, taking Arrow-transparency into account.

quote_arrow_field_deserializer 🔒

This generates code that deserializes a runtime Arrow payload according to the specified datatype.

quote_arrow_field_deserializer_buffer_slice 🔒

This generates code that deserializes a runtime Arrow payload according to the specified datatype.

quote_datatype 🔒

quote_iterator_transparency 🔒

This generates code that maps the data in an iterator in order to apply the Arrow transparency rules to it, if necessary.

should_optimize_buffer_slice_deserialize

Whether or not this object allows for the buffer-slice optimizations.

should_optimize_buffer_slice_deserialize_datatype 🔒

Whether or not this datatype allows for the buffer slice optimizations.