Function re_types_builder::codegen::rust::deserializer::quote_arrow_deserializer_buffer_slice

pub fn quote_arrow_deserializer_buffer_slice(
    arrow_registry: &ArrowRegistry,
    objects: &Objects,
    obj: &Object
) -> TokenStream

This generates code that deserializes a runtime Arrow payload into the specified obj, taking Arrow-transparency into account.

It contains additional performance optimizations based on the inner-type being a non-nullable primitive allowing us to map directly to slices rather than iterating. The ability to use this optimization is determined by should_optimize_buffer_slice_deserialize.

There is a 1:1 relationship between quote_arrow_deserializer_buffer_slice and Loggable::from_arrow:

fn from_arrow(data: &dyn ::arrow2::array::Array) -> DeserializationResult<Vec<Self>> {
    Ok(#quoted_deserializer_)
}

See quote_arrow_deserializer_buffer_slice for additional information.