re_types_builder::codegen::rust::deserializer
Function quote_arrow_field_deserializer

fn quote_arrow_field_deserializer(
    objects: &Objects,
    datatype: &DataType,
    quoted_datatype: &TokenStream,
    is_nullable: bool,
    obj_field_fqname: &str,
    data_src: &Ident,
    inner_repr: InnerRepr,
) -> TokenStream

This generates code that deserializes a runtime Arrow payload according to the specified datatype.

The datatype comes from our compile-time Arrow registry, not from the runtime payload! If the datatype happens to be a struct or union, this will merely inject a runtime call to Loggable::from_arrow_opt and call it a day, preventing code bloat.

data_src is the runtime identifier of the variable holding the Arrow payload (&dyn ::arrow::array::Array). The returned TokenStream always instantiates a Vec<Option<T>>.

This short-circuits on error using the try (?) operator: the outer scope must be one that returns a Result<_, DeserializationError>!