re_types_builder
Module objects

This package implements the semantic pass of the codegen process.

The semantic pass transforms the low-level raw reflection data into higher level types that are much easier to inspect and manipulate / friendler to work with.

Structs

Attributes

A collection of arbitrary attributes.

Object

A high-level representation of a flatbuffers object, which can be either a struct, a union or an enum.

ObjectField

A high-level representation of a flatbuffers field, which can be either a struct member or a union value.

Objects

The result of the semantic pass: an intermediate representation of all available object types; including structs, enums and unions.

ViewReference

Enums

ElementType

The underlying element type for arrays/vectors/maps.

FieldKind

ObjectClass

Is this a struct, enum, or union?

ObjectKind

The kind of the object, as determined by its package root (e.g. rerun.components).

State

Must be set on all archetypes and components

Type

The underlying type of an ObjectField.

Constants

BUILTIN_UNIT_TYPE_FQNAME 🔒

Functions

filepath_from_declaration_file 🔒

is_int 🔒

is_testing_fqname

try_get_enum_fqname 🔒

validate_archetype_field_attributes 🔒

Ensure that each field of an archetype has exactly one of the attr.rerun.component_{required|recommended|optional} attributes.