Module re_types_builder::reflection::reflection

Re-exports

pub use self::bitflags_advanced_features::AdvancedFeatures;

Modules

bitflags_advanced_features 🔒

Structs

BaseType

Enum

EnumArgs

EnumBuilder

EnumVal

EnumValArgs

EnumValBuilder

Field

FieldArgs

FieldBuilder

KeyValue

KeyValueArgs

KeyValueBuilder

Object

ObjectArgs

ObjectBuilder

RPCCall

RPCCallArgs

RPCCallBuilder

Schema

SchemaArgs

SchemaBuilder

SchemaFile

File specific information. Symbols declared within a file may be recovered by iterating over all symbols and examining the declaration_file field.

SchemaFileArgs

SchemaFileBuilder

Service

ServiceArgs

ServiceBuilder

Type

TypeArgs

TypeBuilder

Enums

EnumOffset

EnumValOffset

FieldOffset

KeyValueOffset

ObjectOffset

RPCCallOffset

SchemaFileOffset

SchemaOffset

ServiceOffset

TypeOffset

Constants

ENUM_MAX_BASE_TYPEDeprecated

ENUM_MIN_BASE_TYPEDeprecated

ENUM_VALUES_BASE_TYPEDeprecated

SCHEMA_EXTENSION

SCHEMA_IDENTIFIER

Functions

finish_schema_buffer

finish_size_prefixed_schema_buffer

root_as_schema

Verifies that a buffer of bytes contains a Schema and returns it. Note that verification is still experimental and may not catch every error, or be maximally performant. For the previous, unchecked, behavior use root_as_schema_unchecked.

root_as_schema_unchecked⚠

Assumes, without verification, that a buffer of bytes contains a Schema and returns it.

root_as_schema_with_opts

Verifies, with the given options, that a buffer of bytes contains a Schema and returns it. Note that verification is still experimental and may not catch every error, or be maximally performant. For the previous, unchecked, behavior use root_as_schema_unchecked.

schema_buffer_has_identifier

schema_size_prefixed_buffer_has_identifier

size_prefixed_root_as_schema

Verifies that a buffer of bytes contains a size prefixed Schema and returns it. Note that verification is still experimental and may not catch every error, or be maximally performant. For the previous, unchecked, behavior use size_prefixed_root_as_schema_unchecked.

size_prefixed_root_as_schema_unchecked⚠

Assumes, without verification, that a buffer of bytes contains a size prefixed Schema and returns it.

size_prefixed_root_as_schema_with_opts

Verifies, with the given verifier options, that a buffer of bytes contains a size prefixed Schema and returns it. Note that verification is still experimental and may not catch every error, or be maximally performant. For the previous, unchecked, behavior use root_as_schema_unchecked.